Test fixup
This commit fixed the unit tests such that "make test" passes. It
does not add any new tests, and only modifies TestInitKey to test
initKeyRegistry instead. It does not test any new functionality.

This commit also removes the redundant newKey function since
generatePrivateKeyAndCert does the same job.
