+++ pkg/client/listers/.gitignore
@@ -1,0 +1,24 @@
+# Binaries for programs and plugins
+*.exe
+*.dll
+*.so
+*.dylib
+
+# Test binary, build with `go test -c`
+*.test
+
+# Output of the go coverage tool, specifically when used with LiteIDE
+*.out
+
+# Project-local glide cache, RE: https://github.com/Masterminds/glide/issues/736
+.glide/
+
+/controller
+/kubeseal
+/controller.image
+/*-static
+/controller.yaml
+/sealedsecret-crd.yaml
+/docker/controller
+*.iml
+.idea
